% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertOne.r
\name{convertOne}
\alias{convertOne}
\title{Convert an Excel spreadsheet to match the CoVaxON schema}
\arguments{
\item{file}{Character scalar. Names the path to an Excel file to convert.}
}
\value{
Void
}
\description{
\code{convertOne} reads an Excel file and converts the data
to match the schema required for upload into CoVaxON. The
converted file is saved as CSV and given the same name as
the input Excel file.
}
