% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFiles.r
\name{getFiles}
\alias{getFiles}
\title{List the files to convert}
\usage{
getFiles(excelDir = ".")
}
\arguments{
\item{excelDir}{Character scalar. Names the directory used
to search for .xlsx files. Defaults to the current working directory.}
}
\value{
A \code{list} containing file names. Otherwise, an
error if the directory does not have read access, or does not
contain any \code{.xlsx} files.
}
\description{
\code{getFiles} looks for Excel files within a
directory and returns file names as a list.
For this to work, Excel files must be saved with the
\code{.xlsx} extension.
}
