% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize.r
\name{initialize}
\alias{initialize}
\title{Turn on the log}
\usage{
initialize()
}
\value{
Void
}
\description{
\code{initialize} turns on logging for this package. This function
is designed for internal use and should not be called externally.
}
