% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyErrorFile.r
\name{tidyErrorFile}
\alias{tidyErrorFile}
\title{Tidy the CoVax error file after a Mass Dataload}
\usage{
tidyErrorFile(file, outputFile)
}
\arguments{
\item{file}{Character scalar. Path to the CSV error file returned
by CoVax.}

\item{outputFile}{Character scalar. Path the function should use to write the tidy CSV file.}
}
\value{
\code{tidyErrorFile} returns the input \code{file} invisibly.
}
\description{
\code{tidyErrorFile} iterates through the failed record returned after
a CoVax Mass Dataload and cleans the "Error" column. A CSV is written to
the path specified with \code{outputFile}. \code{tidyErrorFile} takes a
conservative approach to cleaning, given that the "Error" column contains free-text.
}
