% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertMany.r
\name{convertMany}
\alias{convertMany}
\title{Iterate over a list of files and convert each
file to match the CoVax schema}
\usage{
convertMany(path)
}
\description{
Iterate over a list of files and convert each
file to match the CoVax schema
}
