% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertMany.r
\name{convertMany}
\alias{convertMany}
\title{Convert a list of files to match the CoVax schema}
\usage{
convertMany(path, regexp, logFile = "verto2covax.log")
}
\arguments{
\item{path}{Character scalar. Names the directory containing
the Excel files to convert.}
}
\value{
Void
}
\description{
\code{convertMany} iterates over a list of files and calls
\code{\link{convertOne}} on each file. See \code{\link{getFiles}} for
information on how the file list is created.
}
