% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getVaccinationEventId.r
\name{getVaccinationEventId}
\alias{getVaccinationEventId}
\title{Look up the Vaccination Event Id using the Vaccination Event number}
\usage{
getVaccinationEventId(VENumber)
}
\description{
Look up the Vaccination Event Id using the Vaccination Event number
}
